Refactor dijkstra
